Changing RevThreadCtx->RevThread & Moving syscalls to their own file
Doesn't build but redoing changes because of gross merge issues
