+++ src/CMakeLists.txt
@@ -18,7 +18,7 @@ set(RevCPUSrcs
   RevNIC.cc
   RevOpts.cc
   RevProc.cc
-  RevThreadCtx.cc
+  RevThread.cc
   librevcpu.cc
   RevPrefetcher.cc
   RevCoProc.cc
